user role enum validaton + int to bigint for size/price + findByName + location order + params validation

+++ src/entities/property.js
@@ -9,13 +9,13 @@ export class Property{
     @Column('text')
     name;
 
-    @Column('int')
+    @Column('bigint')
     price;
 
     @Column('text')
     description;
     
-    @Column('int')
+    @Column('bigint')
     size;
 
     @Column('text')
